+++ VoiceChat/Runtime/Scripts/VoiceChatSystemBase.cs
@@ -51,8 +51,7 @@ namespace Reflectis.SDK.VoiceChat
         public abstract void MuteAllRemoteStream(bool muteAudio, bool muteVideo);
         public abstract void MuteLocalUser(bool muteAudio);
         public abstract void MuteLocalUser(bool muteAudio, bool muteVideo);
-        public abstract void MuteUser(RemoteUser user);
-        public abstract void EnableUserMic(int userId, bool value);
+        public abstract void EnableUserMic(int actorNumber, bool value);
         public abstract void SetVolume(int volume);
     }
 }